Split RX cycle module
